Replace PdfSharp with PdfPig
