--- src/UmbracoExamine.PDF/PdfSharp/IAdobeGlphListDataProvider.cs
@@ -1,12 +1,0 @@
-﻿using System.IO;
-
-namespace UmbracoExamine.PDF.PdfSharp
-{
-    /// <summary>
-    /// Provide a TextReader for the Adobe Glyph List file.
-    /// </summary>
-    public interface IAdobeGlphListDataProvider
-    {
-        TextReader GetStream();
-    }
-}
