--- src/UmbracoExamine.PDF/PdfSharp/IAdobeGlyphList.cs
@@ -1,7 +1,0 @@
-﻿namespace UmbracoExamine.PDF.PdfSharp
-{
-    public interface IAdobeGlyphList
-    {
-        string Lookup(string glyph);
-    }
-}
