+++ src/UmbracoExamine.PDF/packages.config
@@ -41,6 +41,7 @@
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net472" />
   <package id="NPoco" version="3.9.4" targetFramework="net472" />
   <package id="Owin" version="1.0" targetFramework="net472" />
+  <package id="PdfPig" version="0.1.2" targetFramework="net472" />
   <package id="PDFsharpNetStandard2" version="1.51.4845" targetFramework="net472" />
   <package id="Semver" version="2.0.4" targetFramework="net472" />
   <package id="Serilog" version="2.8.0" targetFramework="net472" />
